Refactor to cleanup disks
