+++ sled-hardware/src/illumos/partitions.rs
