+++ sled-hardware/src/non_illumos/mod.rs
@@ -2,10 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::{
-    Baseboard, DiskError, DiskPaths, DiskVariant, Partition, SledMode,
-    UnparsedDisk,
-};
+use crate::disk::{DiskError, DiskPaths, DiskVariant, Partition, UnparsedDisk};
+use crate::{Baseboard, SledMode};
 use slog::Logger;
 use std::collections::HashSet;
 use tokio::sync::broadcast;
